break out the ci.go file into smaller files
